pyproject.toml
[project]
name = "veritas-rl"
version = "0.1.0"
description = "Veritas RL: Training and evaluation framework for LGC-V2 and Trace environments"
requires-python = ">=3.11, <3.12"
readme = "README.md"
dependencies = [
    # Core ML/AI libraries
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "datasets>=4.2.0",
    "peft>=0.17.1",
    "trl>=0.25.0",
    "huggingface_hub",
    
    # Data processing
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24",
    "scipy>=1.11.0",
    
    # API clients and HTTP
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.11",
    "aiobotocore>=2.23.0",
    
    # Configuration and utilities
    "python-dotenv>=1.0.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.6",
    "jinja2>=3.1.6",
    "click>=8.0.0",
    "click-log>=0.4.0",
    "orjson",
    "aiofiles",
    "tabulate>=0.9.0",
    
    # Logging and monitoring
    "colorlog>=6.7.0",
    "structlog>=23.0.0",
    "wandb>=0.16.0",
    "prometheus-client>=0.21.0",
    "alive_progress>=3.0.0",
    
    # System utilities
    "psutil>=5.9.0",
    "ipykernel>=6.30.1",
    
    # Cloud and storage
    "boto3>=1.34",
    "botocore>=1.38.27",
    "redis==5.*",
    
    # Web framework (for affine-cortex)
    "fastapi>=0.110.1,<0.111",
    "uvicorn==0.22.0",
    "gunicorn>=23.0.0",
    
    # Affine-specific dependencies
    "substrate-interface>=1.7.0",
    "bittensor",
    "bittensor-cli",
    "basilica-sdk>=0.11.0",
    "affinetes @ git+https://github.com/AffineFoundation/affinetes.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["data_processing", "train"]
exclude = [
    "train/verl/**/*",  # Exclude VERL submodule
    "train/parquet/**/*",  # Exclude generated parquet files
    "data_processing/dataset/**/*",  # Exclude generated datasets
    "data_processing/results/**/*",  # Exclude evaluation results
]

